final commit

+++ src/main/java/com/plural/conferencedemo/repo/SessionRepo.java
@@ -3,5 +3,5 @@ package com.plural.conferencedemo.repo;
 import com.plural.conferencedemo.model.Session;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SessionRepo extends JpaRepository<Session, Long> {
+public interface SessionRepo extends JpaRepository<Session, Integer> {
 }
